002-raja-movie-recommendation/requirements.txt
langchain==0.1.0
langgraph==0.0.26
google-generativeai==0.3.2
python-youtube==1.1.0
youtube-search-python==1.6.6
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pandas==2.1.4
streamlit==1.29.0
beautifulsoup4==4.12.2
lxml==4.9.3
